fix: rename history.ts so it stops leaking into the published npm package (fix #8310)
yarn pack unconditionally force-includes any top-level file matching
/+(changes|changelog|history)*, regardless of the package's own "files"
allowlist (packages/excalidraw's package.json only lists "dist/*"). That
made history.ts - full TypeScript source, not the built output - end up
in every published tarball.

Renamed to editorHistory.ts, which the pattern doesn't match. Verified
with an actual `yarn pack`: history.ts is gone from the tarball, only
dist/*, CHANGELOG.md, README.md and package.json remain.
